+++ packages/core-config/src/SecureURL.ts
@@ -1,5 +1,5 @@
 import util, { type InspectOptions } from 'node:util';
-import { sha256 } from './sha256';
+import { SecureString } from './SecureString';
 import { BaseObject, type InspectFunction } from './types';
 
 type UrlObject = {
@@ -13,43 +13,50 @@ type UrlObject = {
  */
 export class SecureURL extends BaseObject {
   readonly #value: URL;
+  readonly #password: SecureString | null;
 
   public get secretValue(): URL {
     return this.#value;
   }
 
-  private constructor(value: URL) {
+  private constructor(value: URL, secret: string | undefined) {
     super();
     this.#value = value;
+    this.#password = SecureString.from(value.password || null, secret);
   }
 
-  public static from(value: undefined | null): null;
-  public static from(value: URL): SecureURL;
-  public static from(value: URL | null | undefined): SecureURL | null {
-    if (value == null) {
-      return null;
+  static factory(secret: string | undefined): (value: URL) => SecureURL {
+    return (value: URL) => SecureURL.from(value, secret);
+  }
+
+  public static from<T extends URL | null | undefined>(value: T, secret?: string): T extends URL ? SecureURL : T {
+    if (value === null) {
+      return null as T extends URL ? SecureURL : T;
     }
-    return new SecureURL(value);
+    if (value === undefined) {
+      return undefined as T extends URL ? SecureURL : T;
+    }
+    return new SecureURL(value, secret) as T extends URL ? SecureURL : T;
   }
 
   override toString(): string {
     const newUrl = new URL(this.#value.href);
-    if (newUrl.password) {
-      newUrl.password = sha256(newUrl.password);
+    if (this.#password !== null) {
+      newUrl.password = this.#password.toString();
     }
     return newUrl.href;
   }
 
-  override toJSON(): object {
+  override toJSON(): UrlObject {
     const newUrl = new URL(this.#value.href);
+    newUrl.password = '';
 
     const searchParams = new URLSearchParams(newUrl.searchParams);
     newUrl.search = '';
 
     let password: string | undefined;
-    if (newUrl.password) {
-      password = sha256(newUrl.password);
-      newUrl.password = '';
+    if (this.#password !== null) {
+      password = this.#password.toString();
     }
 
     const result: UrlObject = {