+++ CHANGELOG.md
@@ -1,1 +1,31 @@
-packages/core-config/CHANGELOG.md
+# Changelog
+
+## [1.0.0] - 2025-01-08
+
+## Added
+
+- `secret` parameter to use `HMAC-SHA256`
+- `createFactory` method to simplify secret creation
+
+## Updates
+
+- Examples in README
+
+## Structure
+
+- Use `packages` and `examples` monorepo structure
+
+## Packaging
+
+- Use `tsup-node`
+
+## Documentation
+
+- Add `readme` example project
+
+## [0.1.0] - 2025-01-07
+
+Initial release.
+
+[1.0.0]: https://github.com/shellicar/core-config/releases/tag/1.0.0
+[0.1.0]: https://github.com/shellicar/core-config/releases/tag/0.1.0